Move document lite forms into views
